requirements.txt
aiogram==3.13.1
asyncpg==0.30.0
python-dotenv==1.0.1
loguru==0.7.2
aiohttp==3.9.3
cachetools==4.2.2
